Simplify unit tests for Hosts. Add one test for environment fallback hosts.
